Moved NSGA classes into algorithms package
